Move utility functions for compression from store.go to utils.go
- Move CompressionTypeFromString() to utils.go
- Move CompressionText to utils.go
- Also move ErrCompressionUnkown to utils.go
