+++ cmd/knoxite/store.go
@@ -12,7 +12,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	humanize "github.com/dustin/go-humanize"
@@ -22,13 +21,13 @@ import (
 	"github.com/spf13/pflag"
 
 	"github.com/knoxite/knoxite"
+	"github.com/knoxite/knoxite/utils"
 )
 
 // Error declarations
 var (
-	ErrRedundancyAmount   = errors.New("failure tolerance can't be equal or higher as the number of storage backends")
 	ErrEncryptionUnknown  = errors.New("unknown encryption format")
-	ErrCompressionUnknown = errors.New("unknown compression format")
+	ErrRedundancyAmount = errors.New("failure tolerance can't be equal or higher as the number of storage backends")
 )
 
 // StoreOptions holds all the options that can be set for the 'store' command
@@ -84,7 +83,7 @@ func store(repository *knoxite.Repository, chunkIndex *knoxite.ChunkIndex, snaps
 	if uint(len(repository.BackendManager().Backends))-opts.FailureTolerance <= 0 {
 		return ErrRedundancyAmount
 	}
-	compression, err := CompressionTypeFromString(opts.Compression)
+	compression, err := utils.CompressionTypeFromString(opts.Compression)
 	if err != nil {
 		return err
 	}
@@ -215,49 +214,6 @@ func executeStore(volumeID string, args []string, opts StoreOptions) error {
 		return err
 	}
 	return repository.Save()
-}
-
-// CompressionTypeFromString returns the compression type from a user-specified string
-func CompressionTypeFromString(s string) (uint16, error) {
-	switch strings.ToLower(s) {
-	case "":
-		// default is none
-		fallthrough
-	case "none":
-		return knoxite.CompressionNone, nil
-	case "flate":
-		return knoxite.CompressionFlate, nil
-	case "gzip":
-		return knoxite.CompressionGZip, nil
-	case "lzma":
-		return knoxite.CompressionLZMA, nil
-	case "zlib":
-		return knoxite.CompressionZlib, nil
-	case "zstd":
-		return knoxite.CompressionZstd, nil
-	}
-
-	return 0, ErrCompressionUnknown
-}
-
-// CompressionText returns a user-friendly string indicating the compression algo that was used
-func CompressionText(enum int) string {
-	switch enum {
-	case knoxite.CompressionNone:
-		return "none"
-	case knoxite.CompressionFlate:
-		return "Flate"
-	case knoxite.CompressionGZip:
-		return "GZip"
-	case knoxite.CompressionLZMA:
-		return "LZMA"
-	case knoxite.CompressionZlib:
-		return "zlib"
-	case knoxite.CompressionZstd:
-		return "zstd"
-	}
-
-	return "unknown"
 }
 
 // EncryptionTypeFromString returns the encryption type from a user-specified string
